requirements.txt
annotated-types==0.6.0
anyio==4.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
distro==1.9.0
exceptiongroup==1.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
llvmlite==0.42.0
more-itertools==10.2.0
numba==0.59.0
numpy==1.26.3
openai==1.7.2
openai-whisper==20231117
PyAudio==0.2.14
pydantic==2.6.1
pydantic_core==2.16.2
pydub==0.25.1
python-dotenv==1.0.0
regex==2023.12.25
requests==2.31.0
six==1.16.0
sniffio==1.3.0
SpeechRecognition==3.10.1
tiktoken==0.6.0
torch==1.11.0
tqdm==4.66.2
typing_extensions==4.9.0
urllib3==2.2.0
websockets==12.0
whisper==1.1.10
